Release supervision and generator
This checks if the latest release of the official runner has a matching
release at this project. If not, a new release is created. When the
release is created, new images are created using the same tag as the
runner SemVer. In the process, as a byproduct, the latest version of
Docker is detected and new images are (re)built, with the same SemVer
tag. These images are used for building the GH runner image