Publish through action
This uses an internal action to publish the containers to their
registries when (auto)releasing and pushing to any branch. The old
tag-based release has been removed.

When releasing, tags of the source Docker and Actions projects are used.
Otherwise, the git sha is used for WIP containers.

All containers are published under the umbrella of the project name at
the GHCR and publication to the DockerHub only happens when a secret is
present.